requirements.txt
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.4
djangorestframework==3.9.0
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pyasn1==0.4.5
Pygments==2.3.1
pytz==2018.7
requests==2.21.0
rsa==4.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
psycopg2==2.7.1
